#0: Cleanup binary op apis
